requirements.txt
psycopg[binary]==3.2.1
python-dateutil==2.9.0.post0
